Implement debug proxy to display error message before install completes
This adds a debug proxy that acts as the debug adapter program before the
coreclr-debug are fully downloaded. If the user tries to start debugging
before download is finished, the proxy displays a sane error message. Once
the download is complete, we rewrite the manifest to no longer call the
proxy.

During the VS Code session that downloads the components, the manifest is
rewritten but not reloaded. In this case the proxy spawns the real
debugger as a child process and proxies its stdin/stdout. Once vs code is
restarted the new manifest is loaded and the proxy is no longer called.

Also adds an empty command that can be run to force activation of the C#
extension, which will kick off debugger acquisition.