Don't package .nyc_output or coverage directories in VSIX
For some reason, on @gregg-miskelly is able to produce much smaller VSIXs that I am. Digging into the folder diffs, it turns out that, on my machine(s), a large amount of code coverage data from one of the node modules (readable-stream) is being included in the VSIX package. This change updates the .vscodeignore file to just ignore those.