Improve error handling for 'isOnPath'
isOnPath calls existsSync, which was not handling ENOTDIR. This can arise
if a portion of a path is not a directory, such as
/dir1/dir2/<executable>/foo/bar. existsSync should return false in these
cases. Additionally, this adds a try/catch to isOnPath that should catch
all other errors from existsSync and treat them as basic failures for that
particular path segment.