Debugger changes for 1.1.4
This updates the clrdbg package to the 25406 build. This includes fixes
for Win7 expression evulation and the System.Task breaking change.
This addresses: https://github.com/OmniSharp/omnisharp-vscode/issues/258

This also the MIEngine to pull in https://github.com/Microsoft/MIEngine/pull/353

Lastly this disables the external console by default. We decided to have it off
by default as we have seen some issues with it on OSX (ex: not always closing
when the target exits), and we are hoping it will be possible soon to use the
new VS Code console window instead. So we decided to turn it off by default, at
least for now.