Implement task to create offline VSIX's
This implements a gulp task that creates platform specific vsix's that
do not require the secondary download/installation steps.

1. Refactor coreclr-debug/main.ts into activation.ts and install.ts
2. Remove VS Code specific modules from install.ts so that it can be
called from gulp
3. Refactor omnisharpDownload.ts so that it has no VS Code dependencies
and can be called from gulp task.
4. Create packaging gulp task that will perform the download of the
debugger and omnisharp in place, package the extension, and clean it up.
This repeats for as many platforms as we need to ship.