Remove System.Diagnostics.StackTrace from debugger install (#620)
The debugger was installing System.Diagnostics.StackTrace which it doesn't
appear to actually need and, in the dev install scenarios at least, was
causing a problem with pulling in the wrong version of
System.Reflection.Metadata. This removes it so the the dev and shipping
code pulls in the same set of assemblies.