Block debugger installation if dotnet cli is too old
This does a version check at the time we verify dotnet is on the path by
running dotnet --info. We use semver to compare the cli version to a hard
coded minimum cli version. If the check fails we use the same UI we
already had for "dotnet not on path" with a different message.