Run tests in VS Code
This change runs unit tests inside of VS Code, making it possible to set
breakpoints and debug test, and to let tests access VS Code services. In
addition, the syntax tests are split out into a separate "npm run
test-syntax" script that runs using Mocha at the command-line. The reason
for this is because the onigurama package used by those tests is
native-compiled, and the compiled binary does not necessarily like running
in VS Code 32-bit on Windows. Ultimately, the syntax tests will be moving
into their own repo, so this is just a temporary measure.