Simplifying remoteProcessPicker (#1123)
* Renaming LaunchConfigurations

To LaunchBrowserPlatformOptions.

* Simplifying retrieving pipeTransport for attach

The launch.json configuration is passed in via the args parameter.
There is no need to read the launch.json file and find the correct
configuration. The old method would only grab the first launch json name
that matched. This change allows duplicate launch names but still grabs
the correct pipeTransport.

* Removing dead code