Linesplit and ohmyzshfix (#1132)
* Renaming LaunchConfigurations

To LaunchBrowserPlatformOptions.

* Simplifying retrieving pipeTransport for attach

The launch.json configuration is passed in via the args parameter.
There is no need to read the launch.json file and find the correct
configuration. The old method would only grab the first launch json name
that matched. This change allows duplicate launch names but still grabs
the correct pipeTransport.

* Removing dead code

* Fixes for OhMyZsh and Line Splitting

Line splitting is dependant on program used. E.g. plink or ssh, so line
splitting for the processes given is handling both \r\n and just \n.

OhMyZsh fails with the current command. Fix is to add extra [] around
existing [].

* Fixing command to work for sh and adding error msg

If a user decides to launch the remoteProcessPicker and clicks outside
of the process picker window, it will try to attach to the process id
'null'. Adding in a new error message.

* Simplifying command