Don't show message about projects having trouble loading when there are only warnings
C# for VS Code is a little over-excited about showing the message that there were problems loading projects. Essentially, it will show it if there were warnings or errors. This changes that behavior to only show it if there are errors because it's perfectly reasonable for a project to contain warnings and not complain to the user every time they open it.