Fix dotnet restore for project.json
In fixing dotnet restore for .csproj, I unintentionally broke it for project.json. The problem here is that OmniSharp returns different information for `Path` with regard to project.json or .csproj. For project.json, it's the directory that the project.json lives in, and for .csproj, it's the actually file path. This change addresses that by adding a `FilePath` property to our `ProjectDescriptor` which represents the real file path and renamed `Path` to `Directory` to be a bit clearer.