Update debugger to 1.12.0 and remove 'runtimeId' support
This checkin updates the debugger to the 1.12.0 release. The major change here is that we are now running on .NET Core 2, and so we can use a single runtime id for all glibc-based Linux distros. This means we can remove all the support we have for runtime id detection and filtering.