Use correct span when requesting code actions from OmniSharp
Fixes #1605

When calling CodeActionProvider.provideCodeActions(...), VS Code passes the range of the word to the left of the editor caret if there's no selection. This ends up causing refactorings like Extract Method to be offered even when the user hasn't selected any code. Now, we'll check to see if there's actually a selection and use that instead of the range that's given.

To fully fix this issue, I've made several changes in OmniSharp: https://github.com/OmniSharp/omnisharp-roslyn/pull/899. Once that goes in, we'll need to update the version of OmniSharp used by C# for VS Code.