Update debugger to 1.12.2 and support large set of Linux distros (#1576)
* Update debugger to 1.12.0 and remove 'runtimeId' support

This checkin updates the debugger to the 1.12.0 release. The major change here is that we are now running on .NET Core 2, and so we can use a single runtime id for all glibc-based Linux distros. This means we can remove all the support we have for runtime id detection and filtering.

* Update debugger to 1.12.2 and update documentation

This updates the debugger to 1.12.2, which among other things, uses .NET Core 2.0-preview2 instead of preview1.