Only send telemetry for Linux distros that appears on an approved list
This change ensures that we only send the name of a Linux distro to the telemetry report if it appears
in an approved list of known distros. The idea here is to avoid sending distro names that we don't recognize
in case the user might have built the distro themselves and included personally-identifiable information in
the distro name.