Add Debug Launch integration test
The integration test:
 - deletes the .vscode directory
 - deletes existing build artifacts
 - issues the `dotnet.generateAssets` command
 - executes the `.NET Core Launch (console)` debug configuration
 - validates that a debugSession becomes active
 - waits for debugging to stop

To enable this, and other integration tests, I exposed an export on the csharp extension which can be awaited until the extension is initialized. Initialized is defined as omnisharp and the debugger being ready for invocation. Without awaiting this promise, commands issues to the extension have unpredictable behaviors since they are registered but not yet able to be handled.