Update debugger packages to 1.14.7
For the last debugger insertion, there was an issue where the OSX package didn't actually contain updated debugger bits. This fixes that problem.

In addition, this also includes the last two months of debugger changes. The primary difference from a C# extension standpoint is that this includes support for embedded source.