Add option to run with a custom mono installation
When `useGlobalMono` is set to `always` or `auto`, the launcher uses whatever
mono it finds in the path, which may or may not be a usable mono version for
omnisharp. The user may have a suitable mono installation for omnisharp
somewhere else, but not set as the default system one, or they may want to
have omnisharp launched with a specific mono for omnisharp for debugging
purposes.

This adds a `monoPath` configuration option, and changes the launcher so that
the environment variables PATH and MONO_GAC_PREFIX are changed to include the
`monoPath` value, if set, when launching mono.