Set QuickPick windows to not lose focus
Enabling ignoreFocusOut allows the QuickPick window to not disappear
when clicking on other windows or when other windows have focus.

An example of this is when a compound is used for launch, we lose focus
and then close the quick pick window as resolved. However, the user has
not selected their process yet and we assume it is undefined.

One catch is that now users have to hit esc to dismiss the process
picker window instead of clicking away.