PR 2540: Remove the bom from the data
This cherry-picks PR 2540 from master. Details of the change:

Fixes: #2521

We were getting the red flame because of a mono behavior, where if we set the encoding to UTF8 then a BOM(Byte Object Mark) is pushed into the stream (output as well as error) as the first character. In the extension we were listening to this input on the error stream and making the flame red.

So the workaround is to remove the "BOM" from the buffer and firing the "STDERR" event only if the length of the BOM is greater than 0.

This will also remove the stray character in the "OmniSharp" log that appears when using mono on Linux/Mac.