Add Error Checking for unsupported platforms
This PR adds in a check to see what platform the extension is running
on. If it is running via Remote SSH, we reccomend users to code on their
host machine and debug remotely. If they are running VS Code somehow on
an unsupported platform, we will mention that the C# extension is
incompatiable with the current platform and arch.