Add vscode implemention of the 'RunTestsInContext' and `DebugTestsInContext' commands
This is the omnisharp-vsocde side of https://github.com/OmniSharp/omnisharp-roslyn/pull/1782. I've implemented support for both of those commands, and bound them to the same default shortcuts as in VS proper. These commands can also be invoked from the command palette.

I've also done a small bit of refactoring to remove references to legacy project.json support. There is no support for loading this project type in Omnisharp anymore, so this was cruft that was unnecessary at this point.