Make updateBuffer use incremental updates
Today, VSCode serializes the entire file whenever it sends a file update request. This can be very costly, particularly for files like Roslyn's NullableReferenceTypeTests.cs, which is 5+ MB. This causes large allocations in both vscode itself and in omnisharp, slowing down typing responsiveness. This commit moves the vscode extension to use the per-line changes that the omnisharp-roslyn host already supports. Editing experience in the afformentioned file is noticeably smoother. Completions are still somewhat slow and work needs to be done on that front, but this makes it possible to actually edit the file with vscode still being responsive.