Support TextEdit in completion responses (#4073)
This is the vscode side of https://github.com/OmniSharp/omnisharp-roslyn/pull/1941. This is written in such a fashion as to be mergeable now, as it will fall back to InsertText if TextEdit is not present in the completion response. After some appropriate period of time, we can remove the InsertText component.