Additional incremental tweaks
My previous incremental change wasn't quite right, and causes files that are not currently opened to never be reloaded. This corrects that change by only skipping file changed on disk events when the file is opened by vscode, meaning that it has an active TextDocument that will be triggered for changes.