Add a SyntaxVisualizer panel
This implements a syntax visualizer view, which is styled very similarly to the panel that ships in the roslyn-sdk for VS. There are 2 parts to this view:

The top tree view shows the entire syntax tree of the current file, including the type of the node and the span of the node. It also uses the same type icons that vscode uses in other contexts to indicate whether the node element is a class or struct. Selecting a node in the tree highlights the span in the current document that corresponds to that node, and clicking in the document reveals the node in the tree view.

The bottom view is another tree view that contains detailed information about the currently-selected node in the top view. It automatically updates any time the top selection is changed, and contains the node type, the syntaxkind, the results of calling GetTypeInfo/GetConversion, GetSymbolInfo, and GetDeclaredSymbol, and all public properties of the node.