Revert incremental change forwarding
While I made this change with good intentions (and the perf benefit when working in Roslyn has been very real), this has proven to be more trouble than it's worth. LSP will be the way to get the perf benefits of incremental changes and should prove to be far more robust. This is a checkout of changeForwarding.ts to d26e995342b8df02135a5cbf0d385a2c02a6dd5b. The only other commits to this file since then have been about 0-based indexes, which no longer matters with this revert.