Filter problematic 'await' keyword suggestions from Razor completion.
The `await` completion item is not compatible with all Razor scenarios.

The `await` completion has been made smarter in that it will now update the containing method signature to include `async`, if it has not been specified.
This is problematic for Razor because it will now suggest making the code behind method that is "wrapping" the use of C# into an async method.
It makes this change by including additional text edits in the completion item.

Example that generates an incompatible completion item:

```
@Da$$
```

Cases where you are in an async method there are no additional text edits and we can continue to offer the `await` keyword.

Example that generates a compatible completion item:

```
@code {
  async Task GetNamesAsync()
  {
    var names = awa$$
  }
}
```