Make the sourceGeneratedDocumentProvider always lazy
GoToDefinition might not immediately go to the document definition for any given location if there are multiple locations (common for source generated symbols), so we shouldn't immediately resolve the generated document during handling of the GoToDefintion response. Instead, only retrieve the file content when the user actually tries to display the file.