Merged PR 457293: Fix trace logging
Previously, trace logging was enabled because vscode automaticallyt ready the microsoft-codeanalysis-languageserver.trace.server option to get if it was enabled.  However when I changed the option to 'dotnet.server.trace' VSCode no longer enabled tracing because there was no option set.

However I just found that vscode has a nice API where we can set the trace level at runtime.  So now we can read our custom option name and pass the value along to vscode.

When logging is set to messages or verbose, you'll be able to see trace logs in `C# LSP trace logs` output window.  Note that these logs are incredibly chatty and verbose which is why they are in a different output window.
![image.png](https://devdiv.visualstudio.com/0bdbc590-a062-4c3f-b0f6-9383f67865ee/_apis/git/repositories/ed74911c-9619-4c14-bd9d-20ecfd2250be/pullRequests/457293/attachments/image.png)