Merged PR 457595: Fix failure to start server on MacOS due to unsigned executable and fix build...
Fix failure to start server on MacOS due to unsigned executable.  Until we figure out how to create Roslyn signed builds signed appropriately for mac (by default it only signs when building on mac) we run using 'dotnet <dll path>' instead.  There is a similar workaround applied in green.

Also fixed build errors on mac by switching to a different way to acquire the package.  Unforuntately the nuget cli on MacOS requires mono to run which is quite a pain.  Nikolche pointed me to this nuget feature which allows me to create essentially an empty csproj and then use the dotnet CLI to download the package instead, which works on all platforms.