Merged PR 464160: Clean and group the configurations
Right now our configuration pages are quite messy. O#'s configurations are put together with blue configurations. Also because now we have `dotnet` prefix configs and `csharp` prefix, the default sorting of VS Code (alphabetical) wouldn't put the `inlay hints` configurations together.

The best way would be VS code gives us the ability to show/hide configurations based on if O# is on/off. However, it is not provided.

The change in this PR:

1. Group all the configurations into three groups.
   a. "C#"
   b. "O#"
    Currently, it is possible to have one layer of grouping in the configurations by adding a different title.
2. In the C# group, use `order` to make `inlay hints` are put together.

TODO:
There are many debug/unit test/razor configurations, need to ask the other team to know if this is only applied in O#

How it looks after:
Grouping:
![image (3).png](https://dev.azure.com/devdiv/0bdbc590-a062-4c3f-b0f6-9383f67865ee/_apis/git/repositories/ed74911c-9619-4c14-bd9d-20ecfd2250be/pullRequests/464160/attachments/image%20%283%29.png)

The shared group (which is inlined by VSCode)
`inlay hints` configs are put together
![image (2).png](https://dev.azure.com/devdiv/0bdbc590-a062-4c3f-b0f6-9383f67865ee/_apis/git/repositories/ed74911c-9619-4c14-bd9d-20ecfd2250be/pullRequests/464160/attachments/image%20%282%29.png)