Merged PR 465893: Remove razor.disabled option
The `razor.disabled` option was introduced in 2018 when the Razor experience in VS Code was still in its infancy. We've since stabilized and the option should no longer be needed. If enabled in Blue, it actually causes the project system to crash with Blazor apps since Roslyn expects Razor to be enabled and tries to send over a dynamic file info request. The easiest solution is to just remove the option entirely.