Merged PR 470357: Add vs-green insertion pipeline
This introduces [a new pipeline](https://devdiv.visualstudio.com/DevDiv/_build?definitionId=18452&_a=summary) that automates inserting 'blue' into 'green' so that green's integration tests run against blue instead of the C# extension found in the marketplace.

See !470347 for the complementary changes made to green for this to be possible.