Pass the service broker pipe name via LSP message rather than a switch
The old approach here meant that we'd always ask for the Dev Kit
pipe name if Dev Kit is installed, which would always trigger a launch
of the Dev Kit server processes, even if we didn't actually have any
reason to use it. The rationale here is if Dev Kit is installed, we
defer to it for the project system loading; if it's not loaded,
then we have no projects, and so this extension won't have anything to
do.