Merged PR 473515: Hide dotnet.generateAssets when DevKit installed
Hide dotnet.generateAssets when DevKit installed

This PR hides away dotnet.generateAssets if the C# DevKit extension is
installed as there are multiple entry points to start debugging.

This also checks to see if we should return undefined or null for
resolveConfigurations. Returning null will hijack the resolve call for
C#DevKit and we want to make sure we give it a chance to resolve for Run
and Debug.