Merged PR 475808: Use environment DOTNET_ROOT for DebugAdapterExe
Use environment DOTNET_ROOT for DebugAdapterExe

This PR updates the DOTNET_ROOT variable for vsdbg/vsdbg-ui if it is set
in the environment. It will prioritize using the environment variable
DOTNET_ROOT over the dotnet CliPath option for omnisharp.

This PR also changes not passing DOTNET_ROOT if either of those are
empty.

Related work items: #1825824