Merged PR 476434: Correct CompletionTriggerKind on '@' Trigger Character
'@' is not a valid trigger character for C#/HTML, so whenever user types '@', Razor would pass to Roslyn 'undefined' for the trigger character with trigger kind being 'Invoked' to get the completion list for when user types '@'. With recent improvements to Razor, the trigger kind that is now being passed when user types '@' is 'TriggerCharacter', which will result in an exception. Set the trigger kind to be 'Invoked' if user types '@' to align with previous behavior.

Related work items: #1828754