Make RoslynLanguageServer.start() no longer async
This method was accidentally doing a fire-and-forget since it didn't
await the underlying launch of the client process. We believe that's
what we want to continue to do, since awaiting the process launch
is dangerous since that can potentially deadlock when you have our
extension calling into Dev Kit or vice versa.