Decouple the log levels for our two output windows
Prior to moving things to LogOutputChannels, we implicitly controlled
the log level of the LSP traces by the log level of our main output
channel: as long as that was debug or trace, we'd also output the LSP
traces. Now each channel behaves independently, with the requirement
that if you want LSP tracing, you must set the trace channel to 'trace'
level. The other settings don't do anything.